Module inexor_rgf_plugin_color::plugin

Structs

ColorPluginImpl

Statics

PLUGIN_CONTEXT

Static plugin context.

PLUGIN_DESCRIPTION

The description of the plugin (CARGO_PKG_DESCRIPTION).

PLUGIN_NAME

The name of the plugin (CARGO_PKG_NAME).

PLUGIN_VERSION

The version of the plugin (CARGO_PKG_VERSION).

Traits

ColorPlugin

Functions

component_manager

component_provider_registry

config_manager

construct_plugin

Constructs the plugin and initializes the plugin context.

entity_behaviour_registry

entity_component_behaviour_registry

entity_instance_manager

entity_type_manager

entity_types_provider_registry

flow_type_manager

flow_types_provider_registry

inject_plugin_context

Returns the static plugin context.

inject_plugin_context_checked

Returns the static plugin context if called after construct_plugin or an empty option if called before construct_plugin.

relation_behaviour_registry

relation_component_behaviour_registry

relation_instance_manager

relation_type_manager

relation_types_provider_registry

web_resource_manager